requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2
    # via flask
flask==1.1.2
    # via -r requirements.in
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via jinja2
mysql-connector-python==8.0.23
    # via -r requirements.in
protobuf==3.15.7
    # via mysql-connector-python
python-dotenv==0.17.0
    # via -r requirements.in
six==1.15.0
    # via protobuf
werkzeug==1.0.1
    # via flask
